Onboard Worker.Core to stylecop
