+++ src/Worker/Core/Logs.cs
@@ -5,7 +5,12 @@ using Microsoft.Extensions.Logging;
 
 namespace Microsoft.DurableTask
 {
-    // NOTE: Trying to make logs consistent with https://github.com/Azure/durabletask/blob/main/src/DurableTask.Core/Logging/LogEvents.cs.
+    /// <summary>
+    /// Log messages.
+    /// </summary>
+    /// <remarks>
+    /// NOTE: Trying to make logs consistent with https://github.com/Azure/durabletask/blob/main/src/DurableTask.Core/Logging/LogEvents.cs.
+    /// </remarks>
     static partial class Logs
     {
         [LoggerMessage(EventId = 55, Level = LogLevel.Information, Message = "{instanceId}: Evaluating custom retry handler for failed '{name}' task. Attempt = {attempt}.")]
