+++ src/Worker/Core/Shims/DurableTaskShimFactory.cs
@@ -21,7 +21,7 @@ public class DurableTaskShimFactory
     readonly ILoggerFactory loggerFactory;
 
     /// <summary>
-    /// Initializes a new instance of <see cref="DurableTaskShimFactory" />.
+    /// Initializes a new instance of the <see cref="DurableTaskShimFactory" /> class.
     /// </summary>
     /// <param name="options">The data converter.</param>
     /// <param name="loggerFactory">The logger factory.</param>
@@ -54,6 +54,8 @@ public class DurableTaskShimFactory
     /// <param name="name">
     /// The name of the activity. This should be the name the activity was invoked with.
     /// </param>
+    /// <typeparam name="TInput">The input type of the activity.</typeparam>
+    /// <typeparam name="TOutput">The output type of the activity.</typeparam>
     /// <param name="implementation">The activity delegate to wrap.</param>
     /// <returns>A new <see cref="TaskActivity" />.</returns>
     public TaskActivity CreateActivity<TInput, TOutput>(
@@ -83,6 +85,8 @@ public class DurableTaskShimFactory
     /// <param name="name">
     /// The name of the orchestration. This should be the name the orchestration was invoked with.
     /// </param>
+    /// <typeparam name="TInput">The input type of the orchestration.</typeparam>
+    /// <typeparam name="TOutput">The output type of the orchestration.</typeparam>
     /// <param name="implementation">The orchestration delegate to wrap.</param>
     /// <param name="parent">The orchestration parent details or <c>null</c> if no parent.</param>
     /// <returns>A new <see cref="TaskOrchestration" />.</returns>
